Array indexing is aweful!
As it currently stands, if the user *does not* specify what indexes
to access, but uses imap/map/reduce without non-static access (e.g.
`x.[0]`), everything works. The issue is two-fold:
 1. Julia indexes start from 1, not 0
 2. Julia is column-major, not row major.

I beleive a solution to this is to modify the AST and not only track
selections into arrays (already encapsulated via `ESel`) but also for
generator expresses for imap. Thus, the following:

    imap { [0] <= iv < [end] }    =>

    imap { [1] <= iv < [end] }