we now correctly traverse and generate arrays
specifically also scalars of array dim-0, though supported in Julia
is a bit tricky to use.

Most examples work now, though there are minor issues with generation
of some of the examples, for instance `letrec` which get passed as
the parameter of a function they call within `in` don't get generated
correctly, leading to the internal functional being called before
the `letrec` itself is defined (see reshape.heh).