fixed further index problem
the `_length` function was initially implemented in a
'drug induced-haze', and so did not provide the length of an array
as given by its shape, but rather the largest component of its shape.

Furthermore, for inline-imap, the index range iterated over was
incorrect, because of both the incorrect `_length` function and the
use of binary operation broadcasts (element-wise operation) to set
IV.

This is further compounded by the fact that Julia prefers linear
over Cartesian indices whereas Heh prefers the latter. Because of this,
the broadcast was actually hiding this flaw by converting the Int scalar
(Julia world, not Heh) into an array.

Both these problems have been fixed and for minor examples seem to
produce the same results as with Heh directly.