resolved indices problem
Previosuly I was using linear indicies which follow the natural order
of a column-major array... this means that for an array[2,2], linear
index 1 = [1,1], and 2 = [2,1] (whereas in Cartesian world this would
be [1,2]). For this reason, generator expressions did not work, or
worked over the wrong range of indices. This commit resolves this by
explicitly using CartesianIndex to generate a correct range expression,
which is wrapped by the heh_setup_range function.